Refactor: move DB state to App.js
